move pkg inside folder to simplify modules
