+++ poutine/pkg/go.mod
@@ -1,11 +1,10 @@
-module github.com/dpatrie/gom-jabbar/poutine
+module github.com/dpatrie/gom-jabbar/poutine/pkg
 
 go 1.15
 
 require (
 	github.com/gomodule/redigo v1.8.2
 	github.com/google/uuid v1.1.2
-	github.com/joho/godotenv v1.3.0
 	github.com/julienschmidt/httprouter v1.3.0
 	github.com/stretchr/testify v1.6.1
 )
