+++ poutine/pkg/robots/bizar.go
@@ -5,8 +5,8 @@ import (
 	"math/rand"
 	"sync"
 
-	"github.com/dpatrie/gom-jabbar/poutine/pubsub"
-	"github.com/dpatrie/gom-jabbar/poutine/resto"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/pubsub"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/resto"
 )
 
 type bizar struct {
