+++ poutine/pkg/robots/montroyashi.go
@@ -4,8 +4,8 @@ import (
 	"math/rand"
 	"strings"
 
-	"github.com/dpatrie/gom-jabbar/poutine/pubsub"
-	"github.com/dpatrie/gom-jabbar/poutine/resto"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/pubsub"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/resto"
 )
 
 type montroyashi struct {
