+++ poutine/pkg/robots/montroyashi_test.go
@@ -4,8 +4,8 @@ import (
 	"testing"
 	"time"
 
-	"github.com/dpatrie/gom-jabbar/poutine/pubsub"
-	"github.com/dpatrie/gom-jabbar/poutine/resto"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/pubsub"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/resto"
 	"github.com/stretchr/testify/assert"
 )
