+++ poutine/pkg/robots/oldoporto.go
@@ -7,8 +7,8 @@ import (
 	"strconv"
 	"sync"
 
-	"github.com/dpatrie/gom-jabbar/poutine/pubsub"
-	"github.com/dpatrie/gom-jabbar/poutine/resto"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/pubsub"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/resto"
 )
 
 type oldoporto struct {
