+++ poutine/pkg/robots/outremona.go
@@ -1,8 +1,8 @@
 package robots
 
 import (
-	"github.com/dpatrie/gom-jabbar/poutine/pubsub"
-	"github.com/dpatrie/gom-jabbar/poutine/resto"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/pubsub"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/resto"
 )
 
 type outremona struct {
