+++ poutine/pkg/robots/verduny.go
@@ -3,8 +3,8 @@ package robots
 import (
 	"time"
 
-	"github.com/dpatrie/gom-jabbar/poutine/pubsub"
-	"github.com/dpatrie/gom-jabbar/poutine/resto"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/pubsub"
+	"github.com/dpatrie/gom-jabbar/poutine/pkg/resto"
 )
 
 type verduny struct {
